Fix for "smart" reader that remembers where a problem happened.
