+++ yagi/src/main/java/org/raml/yagi/framework/nodes/snakeyaml/NodeParser.java
@@ -15,7 +15,6 @@
  */
 package org.raml.yagi.framework.nodes.snakeyaml;
 
-import com.google.common.collect.EvictingQueue;
 import org.raml.v2.api.loader.DefaultResourceLoader;
 import org.raml.v2.api.loader.ResourceLoader;
 import org.raml.yagi.framework.nodes.DefaultPosition;
@@ -25,13 +24,10 @@ import org.yaml.snakeyaml.Yaml;
 import org.yaml.snakeyaml.error.Mark;
 import org.yaml.snakeyaml.error.MarkedYAMLException;
 import org.yaml.snakeyaml.reader.ReaderException;
-import org.yaml.snakeyaml.reader.StreamReader;
 
 import javax.annotation.Nullable;
-import java.io.IOException;
 import java.io.Reader;
 import java.io.StringReader;
-import java.nio.CharBuffer;
 
 import static org.apache.commons.io.IOUtils.LINE_SEPARATOR_UNIX;
 import static org.apache.commons.io.IOUtils.LINE_SEPARATOR_WINDOWS;
@@ -39,138 +35,6 @@ import static org.apache.commons.lang.SystemUtils.IS_OS_WINDOWS;
 
 public class NodeParser
 {
-
-    static class SmartReader extends Reader
-    {
-
-        private final Reader delegate;
-        private char[] lastBufferRead;
-        private int lastBufferSize;
-
-        public SmartReader(Reader delegate)
-        {
-            this.delegate = delegate;
-        }
-
-        @Override
-        public int read(CharBuffer target) throws IOException
-        {
-            return delegate.read(target);
-        }
-
-        @Override
-        public int read() throws IOException
-        {
-            return delegate.read();
-        }
-
-        @Override
-        public int read(char[] cbuf) throws IOException
-        {
-
-            try
-            {
-                int numberOfRead = delegate.read(cbuf);
-                lastBufferRead = cbuf;
-                lastBufferSize = numberOfRead;
-                return numberOfRead;
-            }
-            catch (IOException e)
-            {
-
-                lastBufferSize = 0;
-                throw e;
-            }
-        }
-
-        @Override
-        public int read(char[] cbuf, int off, int len) throws IOException
-        {
-            return delegate.read(cbuf, off, len);
-        }
-
-        @Override
-        public long skip(long n) throws IOException
-        {
-            return delegate.skip(n);
-        }
-
-        @Override
-        public boolean ready() throws IOException
-        {
-            return delegate.ready();
-        }
-
-        @Override
-        public boolean markSupported()
-        {
-            return delegate.markSupported();
-        }
-
-        @Override
-        public void mark(int readAheadLimit) throws IOException
-        {
-            delegate.mark(readAheadLimit);
-        }
-
-        @Override
-        public void reset() throws IOException
-        {
-            delegate.reset();
-        }
-
-        @Override
-        public void close() throws IOException
-        {
-            delegate.close();
-        }
-
-        public String getLastValidString()
-        {
-
-            EvictingQueue<Character> seen = EvictingQueue.create(64);
-
-            for (int i = 0; i < this.lastBufferSize; i++)
-            {
-
-
-                /*
-                 * if ( Character.isISOControl(this.lastBufferRead[i])) {
-                 * 
-                 * seen.add('['); String s = Character.getName(this.lastBufferRead[i]); for (char c : s.toCharArray()) {
-                 * 
-                 * seen.add(c); }
-                 * 
-                 * continue; }
-                 */
-
-                if (!StreamReader.isPrintable(this.lastBufferRead[i]))
-                {
-
-
-                    for (char c : ("[BAD: " + this.lastBufferRead[i] + "]").toCharArray())
-                    {
-
-                        seen.add(c);
-                    }
-
-                    break;
-                }
-
-                seen.add(this.lastBufferRead[i]);
-            }
-
-            Character[] foo = seen.toArray(new Character[0]);
-            StringBuilder sb = new StringBuilder();
-            for (Character character : foo)
-            {
-
-                sb.append(character);
-            }
-
-            return sb.toString();
-        }
-    }
 
     @Nullable
     public static Node parse(ResourceLoader resourceLoader, String resourcePath, Reader reader)
